Added tentative 32 bit support

+++ include/stdlib.h
@@ -2,12 +2,13 @@
 #define STDLIB_HEADER
 
 #include <mcl.h>
+#include <stdint.h>
 #include <stddef.h>
 
-MCL_IMPORT(uint64_t, Kernel32, GetProcessHeap, ());
-MCL_IMPORT(void*, Kernel32, HeapAlloc, (uint64_t, uint32_t, size_t));
-MCL_IMPORT(void*, Kernel32, HeapReAlloc, (uint64_t, uint32_t, void*, size_t));
-MCL_IMPORT(void, Kernel32, HeapFree, (uint64_t, uint32_t, void*));
+MCL_IMPORT(uint32_t, Kernel32, GetProcessHeap, ());
+MCL_IMPORT(void*, Kernel32, HeapAlloc, (uint32_t, uint32_t, size_t));
+MCL_IMPORT(void*, Kernel32, HeapReAlloc, (uint32_t, uint32_t, void*, size_t));
+MCL_IMPORT(void, Kernel32, HeapFree, (uint32_t, uint32_t, void*));
 
 void* malloc(size_t Size) {
 	return HeapAlloc(GetProcessHeap(), 0x8, Size);
